refactor: update project structure to use environmentId instead of projectId across components and API routes; implement environment management features
