+++ apps/dokploy/components/dashboard/search-command.tsx
@@ -20,10 +20,10 @@ import {
 	CommandSeparator,
 } from "@/components/ui/command";
 import { authClient } from "@/lib/auth-client";
-import {
-	extractServices,
-	type Services,
-} from "@/pages/dashboard/project/[projectId]";
+// import {
+// 	extractServices,
+// 	type Services,
+// } from "@/pages/dashboard/project/[projectId]";
 import { api } from "@/utils/api";
 import { StatusTooltip } from "../shared/status-tooltip";
 
@@ -51,7 +51,7 @@ export const SearchCommand = () => {
 
 	return (
 		<div>
-			<CommandDialog open={open} onOpenChange={setOpen}>
+			{/* <CommandDialog open={open} onOpenChange={setOpen}>
 				<CommandInput
 					placeholder={"Search projects or settings"}
 					value={search}
@@ -181,7 +181,7 @@ export const SearchCommand = () => {
 						</CommandItem>
 					</CommandGroup>
 				</CommandList>
-			</CommandDialog>
+			</CommandDialog> */}
 		</div>
 	);
 };
