+++ apps/dokploy/server/api/routers/mariadb.ts
@@ -6,6 +6,7 @@ import {
 	deployMariadb,
 	findBackupsByDbId,
 	findMariadbById,
+	findEnvironmentById,
 	findProjectById,
 	IS_CLOUD,
 	rebuildDatabase,
@@ -41,10 +42,14 @@ export const mariadbRouter = createTRPCRouter({
 		.input(apiCreateMariaDB)
 		.mutation(async ({ input, ctx }) => {
 			try {
+				// Get project from environment
+				const environment = await findEnvironmentById(input.environmentId);
+				const project = await findProjectById(environment.projectId);
+
 				if (ctx.user.role === "member") {
 					await checkServiceAccess(
 						ctx.user.id,
-						input.projectId,
+						project.projectId,
 						ctx.session.activeOrganizationId,
 						"create",
 					);
@@ -56,15 +61,16 @@ export const mariadbRouter = createTRPCRouter({
 						message: "You need to use a server to create a Mariadb",
 					});
 				}
-
-				const project = await findProjectById(input.projectId);
+				
 				if (project.organizationId !== ctx.session.activeOrganizationId) {
 					throw new TRPCError({
 						code: "UNAUTHORIZED",
 						message: "You are not authorized to access this project",
 					});
 				}
-				const newMariadb = await createMariadb(input);
+				const newMariadb = await createMariadb({
+					...input,
+				});
 				if (ctx.user.role === "member") {
 					await addNewService(
 						ctx.user.id,
